projects/5558ba84-c0db-4932-8ad5-686b312dcf4a/requirements.txt
fastapi
sqlite3
pydantic
python-jose
passlib
